Begin cleaning up mdoc format
